More Vendor logic
I have most of it working, the signup sheet takes all the important details that we need and will need to do further frontend logic. My issue now is that I can't seem to save the Vendors to their own database. The signup will succeed and then never show up in the database, but if I go to use the email again it will say "Vendor already exists". So that's something