If the user wants to filter by name then price
I added a catch where if the user selects items they don't want to see then adds a price it doesn't just reset the display.

Though if the user inputs a value that can't be applied to any of the plants then it doesn't change the display
e.g. if the user inputs "10" and the lowest value is "12" then nothing changes